Extract HandleMessages in RChain protocol
