+++ comm/src/main/scala/coop/rchain/comm/rp/Connect.scala
@@ -91,103 +91,5 @@ object Connect {
       _   <- Metrics[F].record("connect-time-ms", tsf - tss)
     } yield ()
 
-  def handlePacket[F[_]: Monad: Time: TransportLayer: ErrorHandler: Log: PacketHandler](
-      remote: PeerNode,
-      maybePacket: Option[Packet]): F[CommunicationResponse] = {
-    val errorMsg = s"Expecting Packet from frame, got something else. Stopping the node."
-    def handleNone: F[CommunicationResponse] =
-      for {
-        _     <- Log[F].error(errorMsg)
-        error = unknownCommError(errorMsg)
-        _     <- errorHandler[F].raiseError[Unit](error)
-      } yield notHandled(error)
-
-    maybePacket.fold(handleNone)(
-      p =>
-        for {
-          local               <- TransportLayer[F].local
-          maybeResponsePacket <- PacketHandler[F].handlePacket(remote, p)
-          maybeResponsePacketMessage = maybeResponsePacket.map(pr =>
-            ProtocolHelper.upstreamMessage(local, AnyProto.pack(pr)))
-        } yield
-          maybeResponsePacketMessage.fold(notHandled(noResponseForRequest))(m =>
-            handledWithMessage(m)))
-  }
-
-  private def handleProtocolHandshake[
-      F[_]: Monad: Time: TransportLayer: NodeDiscovery: Log: ErrorHandler](
-      peer: PeerNode,
-      maybePh: Option[ProtocolHandshake],
-      defaultTimeout: FiniteDuration
-  ): F[CommunicationResponse] =
-    for {
-      local  <- TransportLayer[F].local
-      _      <- getOrError[F, ProtocolHandshake](maybePh, parseError("ProtocolHandshake"))
-      hbrErr <- TransportLayer[F].roundTrip(peer, heartbeat(local), defaultTimeout)
-      commResponse <- hbrErr.fold(
-                       error =>
-                         Log[F]
-                           .warn(
-                             s"Not adding. Could receive Pong message back from $peer, reason: $error")
-                           .as(notHandled(error)),
-                       _ =>
-                         NodeDiscovery[F].addNode(peer) *>
-                           Log[F]
-                             .info(s"Responded to protocol handshake request from $peer")
-                             .as(handledWithMessage(protocolHandshakeResponse(local)))
-                     )
-    } yield commResponse
-
-  private def handleHeartbeat[F[_]: Monad: TransportLayer: ErrorHandler](
-      peer: PeerNode,
-      maybeHeartbeat: Option[Heartbeat]): F[CommunicationResponse] =
-    for {
-      local <- TransportLayer[F].local
-      _     <- getOrError[F, Heartbeat](maybeHeartbeat, parseError("Heartbeat"))
-    } yield handledWithMessage(heartbeatResponse(local))
-
-  def dispatch[
-      F[_]: Monad: Capture: Log: Time: Metrics: TransportLayer: NodeDiscovery: ErrorHandler: PacketHandler](
-      protocol: RoutingProtocol,
-      defaultTimeout: FiniteDuration): F[CommunicationResponse] = {
-
-    def dispatchForUpstream(proto: RoutingProtocol, sender: PeerNode): F[CommunicationResponse] =
-      proto.message.upstream
-        .fold(Log[F].error("Upstream not available").as(notHandled(upstreamNotAvailable))) {
-          usmsg =>
-            usmsg.typeUrl match {
-              // TODO interpolate this string to check if class exists
-
-              case "type.googleapis.com/coop.rchain.comm.protocol.rchain.Heartbeat" =>
-                handleHeartbeat[F](sender, toHeartbeat(proto).toOption)
-
-              case "type.googleapis.com/coop.rchain.comm.protocol.rchain.Packet" =>
-                handlePacket[F](sender, toPacket(proto).toOption)
-
-              case "type.googleapis.com/coop.rchain.comm.protocol.rchain.ProtocolHandshake" =>
-                handleProtocolHandshake[F](sender,
-                                           toProtocolHandshake(proto).toOption,
-                                           defaultTimeout)
-
-              case _ =>
-                Log[F].error(s"Unexpected message type ${usmsg.typeUrl}") *> notHandled(
-                  unexpectedMessage(usmsg.typeUrl)).pure[F]
-            }
-        }
-
-    ProtocolHelper
-      .sender(protocol)
-      .fold(
-        Log[F]
-          .error(s"Sender not present, DROPPING $protocol")
-          .as(notHandled(senderNotAvailable))) { sender =>
-        dispatchForUpstream(protocol, sender)
-      }
-  }
-
-  private def getOrError[F[_]: Applicative: ErrorHandler, A](oa: Option[A],
-                                                             error: CommError): F[A] =
-    oa.fold[F[A]](errorHandler[F].raiseError[A](error))(_.pure[F])
-
   private def errorHandler[F[_]: ErrorHandler]: ErrorHandler[F] = ApplicativeError_[F, CommError]
 }
